alg_v3_aws/requirements.txt
argparse
easydict
h5py
matplotlib
numpy
opencv-python
pyyaml
scipy
tensorboardX
timm==0.4.5 
tqdm
transforms3d
termcolor